fix: tools.devServer type missing some properties
