+++ .changeset/empty-birds-itch.md
@@ -1,0 +1,8 @@
+---
+'@modern-js/core': patch
+'@modern-js/server': patch
+---
+
+fix: tools.devServer type missing some properties
+
+fix: 修复 tools.devServer 类型定义不完整的问题
