+++ packages/cli/core/package.json
@@ -53,6 +53,7 @@
   "devDependencies": {
     "@jest/types": "^27.0.6",
     "@modern-js/babel-preset-app": "workspace:*",
+    "@modern-js/server": "workspace:*",
     "@modern-js/types": "workspace:*",
     "@scripts/build": "workspace:*",
     "@scripts/jest-config": "workspace:*",
